requirements.txt
praw==7.7.1
python-dotenv==1.0.0
pandas==2.0.3
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9
azure-functions==1.15.0
azure-identity==1.14.0
azure-keyvault-secrets==4.9.0